requirements.txt
flask
python-jose
six
openid-connect
pyjwkest
